Finish function to render movies with new schema
