+++ models/movie.ts
@@ -3,36 +3,32 @@ import { Schema, model, models } from "mongoose";
 export interface IMovie {
   title: string;
   description: string;
-  releaseYear: number;
-  genre: string;
+  releaseDate: Date;
+  genres: string[];
   duration: number;
   imgUrl: string;
 
-  screenings: [
-    {
-      displayDate: Date;
-      saloon: string;
-      spokenLang: "en-GB" | "se-SV";
-      subtitLang: "en-GB" | "se-SV" | null;
-    }
-  ];
-
-  reviews: [
-    {
-      reviewerName: string;
-      reviewerText: string;
-      postDate: Date;
-      rating: number;
-    }
-  ];
+  screenings: screening[];
+  reviews: review[];
 }
-
+export type screening = {
+  displayDate: Date;
+  saloon: string;
+  spokenLang: "en-GB" | "se-SV";
+  subtitLang: "en-GB" | "se-SV" | null;
+};
+export type review = {
+  reviewerName: string;
+  reviewerText: string;
+  postDate: Date;
+  rating: number;
+};
 
 const movieSchema = new Schema<IMovie>({
   title: { type: String, required: true },
   description: { type: String, required: true },
-  releaseYear: { type: Number, required: true },
-  genre: { type: String, required: true },
+  releaseDate: { type: Date, required: true },
+  genres: [{ type: Array, required: true }],
   duration: { type: Number, required: true },
   imgUrl: { type: String, required: true },
